@@ -3,9 +3,9 @@ from setuptools import find_packages, setup
 setup(
     name="deepthinking",
     version="0.1.0",
-    description="Deep Thinking Neural Networks Implementation",
-    author="Your Name",
+    description="Deep Thinking",
+    author="Elisa Klunder",
     packages=find_packages(),
     install_requires=[],
-    python_requires=">=3.7",
+    python_requires="==3.10.17",
 )
